fix(ken-552): draw before the read, keep the breaks [no-changelog]
Three defects this branch armed.

A break the caller wrote is structure and now survives. Escaping the whole
message flattened it, so kendex diff's leading newline before every file
heading reached the reader as the two characters a backslash and an n, and
marketplace author's multi-line error collapsed onto one line. The ui entry
points are line-oriented: one split, shared by escaped() and tell(), each
line escaped on its own. Covered by a unit test over both halves, since a
fix for either one alone regresses the other.

No read can happen over a block that is still open. repo_effects::confirm
moves onto ui::confirm, whose plain rendering is the same bytes it wrote
itself; harness_picker's raw write and read moves into ui::ask. Every stdin
read in the CLI is now inside ui/prompt.rs and both readers flush first, so
the ordering holds by construction rather than by remembering.

kendex add of a collection closes on the same ledger add gives a package.
Each step hands back the counts it already took and the closing line is
their sum, emitted after the repository-effects walkthrough.

crates/cli/src/ui.rs crossed the ratchet's 400-line threshold as a new
offender, so the block a run is in the middle of saying, and the frame
around it, move to ui/blocks.rs. No baseline row was added.
